# Optimize Windows container statistics