@dcantah 2022-04-20
> I like the idea. Can we not just query the working set size associated to the JO handle for the container in one pass? It seems like that should be a feature if it isnt already :)

At least not that I could find, I've asked some peeps to look around if there's something hidden..