@kevpar 2022-05-05
> > I think these changes go beyond "optimizing". For clarify, we should phrase the commit differently so the impact is clear. I'd suggest something like "wcow-process: Query stats directly from shim".
> 
> @kevpar I can change the squashed commit msg on check-in to be more clear here

Can we change the PR title to? I think that's included in the squash commit title.